// apps/dashboard/@types/complaintTypes.types.ts
export type complaintTypes = {
  _id: string;
  phone: string;
  userName: string;
  userId: string;
  supervisorId?: string;
  complaintType: string;
  complaintAddress: string;
  complaintDes?: string;
  wsscStatement?: string;
  ImageUrl?: string;
  feedback?: any;
  response?: any;
  VideoUrl?: string;
  status: any;
  createdAt: string;
};